Road to 3.00
Nue 3.00
- Change CDN Assets
- A lil bit re-designed.
- Reorganize Default Menu
- Implementing PJAX
- Adding System Information page as an additional default installation
- Adding readable member's password for Administrator
- Change Iconify to Bootstrap Icons for efficiency
- Change default roles & permissions
- Selectable language between Indonesian & English

+++ resources/views/settings/menu/index.blade.php
@@ -1,9 +1,13 @@
-@extends('layouts.app')
+@extends('layouts.base')
 @section('title', $title)
 
 @section('css')
+	<link rel="stylesheet" href="{{ config('nue.brand.cdn') }}/vendor/tom-select/css/tom-select.bootstrap5.min.css">
 	<link rel="stylesheet" href="//cdn.jsdelivr.net/npm/nestable2@1.6.0/jquery.nestable.min.css">
 	<style>
+		.dd {
+			max-width: none;
+		}
 		.dd-handle {
 			height: 40px;
 			margin: 5px;
@@ -11,12 +15,8 @@
 			border: 1px solid #eeeff5;
 		}
 		.dd-list .dd-list {
-			padding-left: 40px;
+			padding-left: 25px;
 		}
-		.tom-select-custom .ts-control.multi .ts-input>div {
-			padding: 0.25rem 0.5rem!important;
-		}
-
 		.swal2-modal .swal2-title {
 			font-size: 20px;
 		}
@@ -33,10 +33,12 @@
 @endsection
 
 @section('js')
+	<script src="{{ config('nue.brand.cdn') }}/vendor/tom-select/js/tom-select.complete.min.js"></script>
+	<script src="{{ config('nue.brand.cdn') }}/js/nue-tom-select.js"></script>
 	<script type="text/javascript" src="//cdn.jsdelivr.net/npm/nestable2@1.6.0/jquery.nestable.min.js"></script>
 	<script>
 		$(document).ready(function() {
-			Nue.components.NueSelect.init('.js-select')
+			Nue.components.NueTomSelect.init('.js-select');
 
 			$('#nestable-menu').on('click', function(e) {
 				var target = $(e.target),
@@ -65,9 +67,9 @@
 
 		$('.menu-delete').click(function() {
 			var id = $(this).data('id');
-			swal({
+			Swal.fire({
 				title: "Are you sure ?",
-				type: "warning",
+				icon: "warning",
 				text: "We will delete this menu permanently.",
 				width: "300px", 
 				showCancelButton: true,
@@ -75,27 +77,29 @@
 				confirmButtonText: "Confirm",
 				cancelButtonText: "Cancel",
 				allowOutsideClick: false
-			}).then(function(data) {
-				$.ajax({
-					type: 'DELETE',
-					url: '{{ route("$prefix.index") }}' + '/' + id,
-					data: {
-						"id": id,
-						"_token": $('meta[name="csrf-token"]').attr('content'),
-					},
-					success: function(data) {
-						if(data == "success") {
-							history.go(0);
-						} else {
-							swal({
-								type: 'warning',
-								title: 'Galat!',
-								width: "300px", 
-								html: 'Terjadi kesalahan saat menghapus data. Coba sekali lagi.'
-							});
+			}).then(function(result) {
+				if(result.value) {
+					$.ajax({
+						type: 'DELETE',
+						url: '{{ route("$prefix.index") }}' + '/' + id,
+						data: {
+							"id": id,
+							"_token": $('meta[name="csrf-token"]').attr('content'),
+						},
+						success: function(data) {
+							if(data == "success") {
+								history.go(0);
+							} else {
+								Swal.fire({
+									icon: 'warning',
+									title: 'Galat!',
+									width: "300px", 
+									html: 'Terjadi kesalahan saat menghapus data. Coba sekali lagi.'
+								});
+							}
 						}
-					}
-				});
+					});
+				}
 			});
 		});
 	</script>
@@ -104,58 +108,55 @@
 @section('content')
 
     @include('nue::partials.breadcrumb', ['lists' => [
-        'Settings' => 'javascript:;', 
+        'Nue Settings' => 'javascript:;', 
         $title => 'active'
     ]])
 
-    @include('nue::partials.datatable.header', [
-        'title' => $title, 
-        'description' => 'Here is a list of all your data from your database.'
-    ])
-
-    <div class="content container-fluid p-3">
-    	<div class="row">
-    		<div class="col-md-7 pe-2">
-    			<div class="card rounded-1 mb-3">
-    				<div class="card-header">
+    <div class="content container-fluid bg-light h-breadcrumb p-2">
+    	<div class="row gx-2">
+    		<div class="col-md-7">
+    			<div class="card shadow-none border rounded-0 mb-3">
+    				<div class="card-header bg-light p-2">
     					{!! Form::open(['route' => "$prefix.store"]) !!}
                             <div id="nestable-menu" class="btn-group">
-								<a class="btn btn-primary btn-sm" data-action="expand-all" title="Expand">
-									Expand
+								<a class="btn btn-primary btn-sm" data-action="expand-all" title="{{ __('Expand') }}">
+									{{ __('Expand') }}
 								</a>
-								<a class="btn btn-primary btn-sm" data-action="collapse-all" title="Collapse">
-									Collapse
+								<a class="btn btn-primary btn-sm" data-action="collapse-all" title="{{ __('Collapse') }}">
+									{{ __('Collapse') }}
 								</a>
 							</div>
 							<div class="btn-group">
-								<button type="submit" class="btn btn-info btn-sm" title="Save">
-									<span class="hidden-xs">&nbsp;Save</span>
-								</button>
-								<a class="btn btn-white btn-sm" title="Refresh" onclick="return history.go(0)">
-									<span class="hidden-xs">&nbsp;Refresh</span>
+								<a class="btn btn-white btn-sm" title="{{ __('Refresh') }}" onclick="return history.go(0)">
+									<span class="hidden-xs">{{ __('Refresh') }}</span>
 								</a>
+								<button type="submit" class="btn btn-info btn-sm" title="{{ __('Save') }}">
+									<span class="hidden-xs">{{ __('Save') }}</span>
+								</button>
 							</div>
                             {!! Form::hidden('order', null, ['id' => 'nestable-output']) !!}
                         {!! Form::close() !!}
     				</div>
-    				<div class="card-body">
+    				<div class="card-body p-2">
     					<div class="dd" id="nestable">
 							<ol class="dd-list">
 								@foreach(config('nue.database.menu_model')::whereParentId(0)->orderBy('order')->get() as $i => $temp)
 									<li class="dd-item" data-id="{!! $temp->id !!}">
 										<div class="dd-handle d-flex justify-content-between">
 											<div>
-												<span class="iconify me-1" data-icon="{!! $temp->icon !!}"></span>
-												{!! $temp->title !!}
-												<span class="text-muted">{!! $temp->uri !!}</span>
+												@if(!is_null($temp->icon))
+													<i class="bi bi-{!! $temp->icon !!} me-1"></i>
+												@endif
+												{!! __($temp->title) !!}
+												<span class="text-muted ms-1">{!! $temp->uri !!}</span>
 											</div>
 											<div class="dd-nodrag">
 												<a href="{!! route("$prefix.edit", $temp->id) !!}" class="text-success me-1">
-													<span class="iconify i-lg" data-icon="icon-park-twotone:edit"></span> 
-													Edit
+													<i class="bi bi-pencil-square"></i> 
+													{{ __('Edit') }}
 												</a>
 												<a href="javascript:void(0);" class="menu-delete text-danger" data-id="{!! $temp->id !!}">
-													<span class="iconify i-lg" data-icon="icon-park-twotone:delete-five"></span>
+													<i class="bi bi-trash-fill"></i>
 												</a>
 											</div>
 										</div>
@@ -165,17 +166,19 @@
 													<li class="dd-item" data-id="{!! $side->id !!}">
 														<div class="dd-handle d-flex justify-content-between">
 															<div>
-																<span class="iconify me-1" data-icon="{!! $side->icon !!}"></span>
-																{!! $side->title !!}
-																<span class="text-muted">{!! $side->uri !!}</span>
+																@if(!is_null($side->icon))
+																	<i class="bi bi-{!! $side->icon !!} me-1"></i>
+																@endif
+																{!! __($side->title) !!}
+																<span class="text-muted ms-1">{!! $side->uri !!}</span>
 															</div>
 															<div class="dd-nodrag">
 																<a href="{!! route("$prefix.edit", $side->id) !!}" class="text-success me-1">
-																	<span class="iconify i-lg" data-icon="icon-park-twotone:edit"></span> 
-																	Edit
+																	<i class="bi bi-pencil-square"></i>
+																	{{ __('Edit') }}
 																</a>
 																<a href="javascript:void(0);" class="menu-delete text-danger" data-id="{!! $side->id !!}">
-																	<span class="iconify i-lg" data-icon="icon-park-twotone:delete-five"></span>
+																	<i class="bi bi-trash-fill"></i>
 																</a>
 															</div>
 														</div>
@@ -190,18 +193,17 @@
     				</div>
     			</div>
     		</div>
-    		<div class="col-md-5 ps-2">
+    		<div class="col-md-5">
     			{!! Form::open(['route' => "$prefix.store"]) !!}
-	    			<div class="card rounded-1">
-	    				<div class="card-header with-border">
-	    					<h3 class="card-title mb-0">New</h3>
+	    			<div class="card shadow-none border rounded-0">
+	    				<div class="card-header bg-light p-3">
+	    					<h4 class="card-title mb-0">
+	    						{{ __('New') }}
+	    					</h4>
 	    				</div>
 						@include("$view.form")
 						<div class="card-footer p-3">
-							<div class="d-flex justify-content-end">
-								<button type="reset" class="btn btn-sm btn-warning me-2">Reset</button>
-								<button type="submit" class="btn btn-sm btn-info">Submit</button>
-							</div>
+							<button type="submit" class="btn btn-sm btn-info">{{ __('Submit') }}</button>
 						</div>
 					</div>
 					{!! Form::hidden('purpose', 'add') !!}
